src/backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "AIfoundry.app backend API"
authors = [
    {name = "Govind", email = "gok@microsoft.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.7",
    "azure-ai-projects>=1.0.0b5",
    "azure-core>=1.30.0",
    "azure-cosmos>=4.7.0",
    "openai>=1.0.0",
    "mcp>=1.11.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.1.1",
    "pyjwt>=2.10.1",
    "pynacl>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
